feature: added reducer, but it is still a work in progress
