+++ src/context/ShoppingContext.js
@@ -1,7 +1,7 @@
 import { createContext } from "react";
 
 const ShoppingContext = createContext({
-  name: "",
+  name: {},
   email: "",
   phoneNumber: "",
   submitStep1: () => {},
