+++ src/pages/Checkout/Step2.js
@@ -15,11 +15,7 @@ function Step2({ cartItems, handleRemove, handleChange, ...routeProps }) {
         <FormShopping {...routeProps} />
       </div>
       <div className="col">
-        <OrderSummary
-          cartItems={cartItems}
-          handleRemove={handleRemove}
-          handleChange={handleChange}
-        />
+        <OrderSummary cartItems={cartItems} />
       </div>
     </div>
   );
